Add unit testing for IGEMM QAI8
There are some TODOs left in code. They should be addressed, but the
change is useful as is.

There is also an issue where the RHS packing doesn't seem to be working.
This is worked around by always packing entire RHS, and not only the
part that is needed for the output portion. This need to be investigated
before releasing to the wild

Signed-off-by: Emil Ohlsson <emil.ohlsson@arm.com>

Reviewed-by: Emil Ohlsson <emil.ohlsson@arm.com>
Approved-by: Jakub Sujak <jakub.sujak@arm.com>